requirements.txt
ollama>=0.5.0
pydantic>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
opencv-python>=4.8.0
